Add files via upload
A megabyte is 8000 kilobits. Write a program that will read in the bit rate in kilo-bits per second and the duration in minutes of a video clip and will output the file size of the video clip in megabytes.
Define a function to compute the file size of the video clip. Your program should use a globally defined constant for the number of kilobits per megabyte.